Kernel: Fix physical address size for x86_32
Having 32 bit address space does not mean physical address space
is also only 32 bits...

+++ kernel/include/kernel/Memory/Types.h
@@ -21,6 +21,6 @@ namespace Kernel
 {
 
 	using vaddr_t = uintptr_t;
-	using paddr_t = uintptr_t;
+	using paddr_t = uint64_t;
 
 }
